requirements.txt
flask
gunicorn
africastalking
python-dotenv